+++ changelog.md
@@ -4,3 +4,9 @@
 - criacao do projeto, configurando cores/fontes e montagem das rotas
 - Montagem do footer e criando rota do checkout
 - Micro componentes - Button, Input e Icons
+
+
+[09-02-2025]
+- criacao do product card
+- criacao de top bar - falta animacao
+- criacao do header